sprite-server/api/requirements.txt
fastapi==0.115.6
uvicorn==0.32.1
pillow==11.0.0
python-multipart==0.0.20
